pyproject.toml
[tool.poetry]
name = "electrosales-net"
version = "0.1.0"
description = "TechStack for electronical sales network app"
authors = ["Edenerus <markeev.25@mail.ru>"]
readme = "README.md"
packages = [{include = "electrosales_net"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
django-filter = "^23.2"
docker = "^6.0.1"
django-environ = "^0.10.0"
pytest = "^7.3.1"
django-rest-swagger = "^2.2.0"
drf-spectacular = "^0.26.2"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
